Add basic ads tracking support
This implementation is based on https://github.com/newrelic/video-jwplayer-js
to add ads support separated from the caf tracker.

I added an instance variable playerManager to prevent the repeating
of cast.framework API.

I refactored the debug core events in the tracker to be implemented
using the builtin debugCommonVideoEvents function.